requirements.txt
asgiref==3.5.0
autopep8==1.6.0
bootstrap-admin==0.4.4
click==8.1.2
colorama==0.4.4
dj-database-url==0.5.0
Django==4.0.3
django-admin-interface==0.19.1
django-ckeditor==6.4.2
django-colorfield==0.7.1
django-countries==7.3.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-grappelli==3.0.3
django-heroku==0.3.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-phone-field==1.8.1
django-phonenumber-field==6.3.0
django-quill==0.1.0
django-tinymce==3.4.0
djlint==0.7.6
gunicorn==20.1.0
importlib-metadata==4.11.3
pathspec==0.9.0
phonenumbers==8.12.52
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
PyYAML==6.0
regex==2022.3.15
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomlkit==0.9.2
tqdm==4.64.0
typing_extensions==4.3.0
tzdata==2021.5
whitenoise==6.2.0
zipp==3.8.0
